+++ MatrixKit/Controllers/MXKAuthenticationViewController.h
@@ -50,15 +50,11 @@ extern NSString *const MXKAuthErrorDomain;
 @interface MXKAuthenticationViewController : MXKViewController <UITextFieldDelegate, MXKAuthInputsViewDelegate>
 {
 @protected
-    /**
-     Array of flows supported by the home server and implemented by the view controller (for the current auth type).
-     */
-    NSMutableArray *supportedFlows;
     
     /**
-     The current selected login flow
+     Reference to any opened alert view.
      */
-    MXLoginFlow *selectedFlow;
+    MXKAlert *alert;
 }
 
 @property (weak, nonatomic) IBOutlet UIImageView *welcomeImageView;
@@ -99,9 +95,9 @@ extern NSString *const MXKAuthErrorDomain;
 @property (nonatomic) MXKAuthenticationType authType;
 
 /**
- The current selected login flow
+ The current view in which authentication inputs are displayed (`MXKAuthInputsView-inherited` instance).
  */
-@property (nonatomic) MXLoginFlow *selectedFlow;
+@property (nonatomic) MXKAuthInputsView *authInputsView;
 
 /**
  The default home server url (nil by default).
@@ -114,51 +110,57 @@ extern NSString *const MXKAuthErrorDomain;
 @property (nonatomic) NSString *defaultIdentityServerUrl;
 
 /**
+ Enable/disable overall the user interaction option.
+ It is used during authentication process to prevent multiple requests.
+ */
+@property(nonatomic,getter=isUserInteractionEnabled) BOOL userInteractionEnabled;
+
+/**
  The delegate for the view controller.
  */
 @property (nonatomic) id<MXKAuthenticationViewControllerDelegate> delegate;
 
 /**
- *  Returns the `UINib` object initialized for a `MXKAuthenticationViewController`.
- *
- *  @return The initialized `UINib` object or `nil` if there were errors during initialization
- *  or the nib file could not be located.
- *
- *  @discussion You may override this method to provide a customized nib. If you do,
- *  you should also override `authenticationViewController` to return your
- *  view controller loaded from your custom nib.
+ Returns the `UINib` object initialized for a `MXKAuthenticationViewController`.
+ 
+ @return The initialized `UINib` object or `nil` if there were errors during initialization
+ or the nib file could not be located.
+ 
+ @discussion You may override this method to provide a customized nib. If you do,
+ you should also override `authenticationViewController` to return your
+ view controller loaded from your custom nib.
  */
 + (UINib *)nib;
 
 /**
- *  Creates and returns a new `MXKAuthenticationViewController` object.
- *
- *  @discussion This is the designated initializer for programmatic instantiation.
- *
- *  @return An initialized `MXKAuthenticationViewController` object if successful, `nil` otherwise.
+ Creates and returns a new `MXKAuthenticationViewController` object.
+ 
+ @discussion This is the designated initializer for programmatic instantiation.
+ 
+ @return An initialized `MXKAuthenticationViewController` object if successful, `nil` otherwise.
  */
 + (instancetype)authenticationViewController;
 
 /**
- Register the MXKAuthInputsView class that will be used to display inputs for the designated flow and authentication type.
+ Register the MXKAuthInputsView class that will be used to display inputs for an authentication type.
  
- By default only 'MXKAuthInputsPasswordBasedView' class is registered for 'kMXLoginFlowTypePassword' flow and 'MXKAuthenticationTypeLogin' authentication.
+ By default the 'MXKAuthInputsPasswordBasedView' class is registered for 'MXKAuthenticationTypeLogin' authentication.
+ No class is registered for 'MXKAuthenticationTypeRegister' type.
  
- @param authInputsViewClass a MXKAuthInputsView-inherited class that will be used for the designated flow.
- @param flowType the concerned flow type.
+ @param authInputsViewClass a MXKAuthInputsView-inherited class.
  @param authType the concerned authentication type
  */
-- (void)registerAuthInputsViewClass:(Class)authInputsViewClass forFlowType:(MXLoginFlowType)flowType andAuthType:(MXKAuthenticationType)authType;
+- (void)registerAuthInputsViewClass:(Class)authInputsViewClass forAuthType:(MXKAuthenticationType)authType;
 
 /**
- Check login mechanism supported by the server and the application.
+ Refresh login/register mechanism supported by the server and the application.
  */
-- (void)refreshSupportedAuthFlow;
+- (void)refreshAuthenticationSession;
 
 /**
- Handle supported flows returned by the server.
+ Handle supported flows and associated information returned by the home server.
  */
-- (void)handleHomeServerFlows:(NSArray *)flows;
+- (void)handleAuthenticationSession:(MXAuthenticationSession *)authSession;
 
 /**
  Customize the MXHTTPClientOnUnrecognizedCertificate block that will be used to handle unrecognized certificate observed during authentication challenge from a server.
@@ -174,6 +176,27 @@ extern NSString *const MXKAuthErrorDomain;
  - 'UIControlEventValueChanged' for each UISwitch instance.
  */
 - (IBAction)onButtonPressed:(id)sender;
+
+/**
+ Set the home server url and force a new authentication session.
+ The default home server url is used when the provided url is nil.
+ 
+ @param homeServerUrl the home server url to use
+ */
+- (void)setHomeServerTextFieldText:(NSString *)homeServerUrl;
+
+/**
+ Set the identity server url.
+ The default identity server url is used when the provided url is nil.
+ 
+ @param identityServerUrl the identity server url to use
+ */
+- (void)setIdentityServerTextFieldText:(NSString *)identityServerUrl;
+
+/**
+ Force dismiss keyboard
+ */
+- (void)dismissKeyboard;
 
 @end