+++ MatrixKit/Models/Account/MXKAccount.h
@@ -61,6 +61,12 @@ typedef BOOL (^MXKAccountOnCertificateChange)(MXKAccount *mxAccount, NSData *cer
 @property (nonatomic) NSString *identityServerURL;
 
 /**
+ The Push Gateway URL used to send event notifications to (nil by default).
+ This URL should be over HTTPS and never over HTTP.
+ */
+@property (nonatomic) NSString *pushGatewayURL;
+
+/**
  The matrix REST client used to make matrix API requests.
  */
 @property (nonatomic, readonly) MXRestClient *mxRestClient;
@@ -86,6 +92,18 @@ typedef BOOL (^MXKAccountOnCertificateChange)(MXKAccount *mxAccount, NSData *cer
  The account display name based on user id and user displayname (if any).
  */
 @property (nonatomic, readonly) NSString *fullDisplayName;
+
+/**
+ The 3PIDs linked to this account.
+ [self load3PIDs] must be called to update the property.
+ */
+@property (nonatomic, readonly) NSArray<MXThirdPartyIdentifier *> *threePIDs;
+
+/**
+ The email addresses linked to this account.
+ This is a subset of self.threePIDs.
+ */
+@property (nonatomic, readonly) NSArray<NSString *> *linkedEmails;
 
 /**
  The account user's presence (`MXPresenceUnknown` by default, available if matrix session `mxSession` is opened).
@@ -233,6 +251,15 @@ typedef BOOL (^MXKAccountOnCertificateChange)(MXKAccount *mxAccount, NSData *cer
  @param failure A block object called when the operation fails.
  */
 - (void)changePassword:(NSString*)oldPassword with:(NSString*)newPassword success:(void (^)())success failure:(void (^)(NSError *error))failure;
+
+/**
+ Load the 3PIDs linked to this account.
+ This method must be called to refresh self.threePIDs and self.linkedEmails.
+
+ @param success A block object called when the operation succeeds.
+ @param failure A block object called when the operation fails.
+ */
+- (void)load3PIDs:(void (^)())success failure:(void (^)(NSError *error))failure;
 
 #pragma mark - Push notification listeners
 /**