+++ MatrixKit/Models/MXK3PID.h
@@ -38,6 +38,16 @@ typedef enum : NSUInteger {
 @property (nonatomic, readonly) NSString *address;
 
 /**
+ The current client secret key used during email validation.
+ */
+@property (nonatomic, readonly) NSString *clientSecret;
+
+/**
+ The current session identifier during email validation.
+ */
+@property (nonatomic, readonly) NSString *sid;
+
+/**
  The id of the user on Matrix.
  nil if unknown or not yet resolved.
  */
@@ -57,8 +67,9 @@ typedef enum : NSUInteger {
 /**
  Start the validation process 
  The identity server will send a validation token to the user's address.
- This validation token must be then send back to the identity server with [MXK3PID validateWithToken]
- in order to complete the 3PID authentication.
+ 
+ In case of email, the end user must click on the link in the received email
+ to validate their email address in order to be able to call add3PIDToUser successfully.
  
  @param restClient used to make matrix API requests during validation process.
  @param success A block object called when the operation succeeds.
@@ -67,29 +78,17 @@ typedef enum : NSUInteger {
 - (void)requestValidationTokenWithMatrixRestClient:(MXRestClient*)restClient
                                            success:(void (^)())success
                                            failure:(void (^)(NSError *error))failure;
-
 /**
- Complete the 3rd party id validation by sending the validation token the user received.
+ Link a 3rd party id to the user.
  
- @param validationToken the validation token the user received.
- @param success A block object called when the operation succeeds. It indicates if the
- validation has succeeded.
- @param failure A block object called when the operation fails.
- */
-- (void)validateWithToken:(NSString*)validationToken
-              success:(void (^)(BOOL success))success
-              failure:(void (^)(NSError *error))failure;
-
-/**
- Link an authenticated 3rd party id to a Matrix user id.
- 
- @param userId the Matrix user id to link the 3PID with.
+ @param bind whether the homeserver should also bind this third party identifier
+        to the account's Matrix ID with the identity server.
  @param success A block object called when the operation succeeds. It provides the raw
  server response.
  @param failure A block object called when the operation fails.
  */
-- (void)bindWithUserId:(NSString*)userId
-         success:(void (^)())success
-         failure:(void (^)(NSError *error))failure;
+- (void)add3PIDToUser:(BOOL)bind
+              success:(void (^)())success
+              failure:(void (^)(NSError *error))failure;
 
 @end