+++ MatrixKit/Views/Authentication/MXKAuthInputsEmailCodeBasedView.m
@@ -19,7 +19,6 @@
 #import "NSBundle+MatrixKit.h"
 
 @implementation MXKAuthInputsEmailCodeBasedView
-@dynamic displayNameTextField;
 
 + (UINib *)nib
 {
@@ -34,48 +33,74 @@
     _userLoginTextField.placeholder = [NSBundle mxk_localizedStringForKey:@"login_user_id_placeholder"];
     _emailAndTokenTextField.placeholder = [NSBundle mxk_localizedStringForKey:@"login_email_placeholder"];
     _promptEmailTokenLabel.text = [NSBundle mxk_localizedStringForKey:@"login_prompt_email_token"];
+    
+    _displayNameTextField.placeholder = [NSBundle mxk_localizedStringForKey:@"login_display_name_placeholder"];
 }
 
-- (CGFloat)actualHeight
+#pragma mark -
+
+- (BOOL)setAuthSession:(MXAuthenticationSession *)authSession withAuthType:(MXKAuthenticationType)authType;
 {
-    if (self.authType == MXKAuthenticationTypeLogin)
+    // Validate first the provided session
+    MXAuthenticationSession *validSession = [self validateAuthenticationSession:authSession];
+    
+    if ([super setAuthSession:validSession withAuthType:authType])
     {
-        return self.displayNameTextField.frame.origin.y;
+        // Set initial layout
+        self.userLoginTextField.hidden = NO;
+        self.promptEmailTokenLabel.hidden = YES;
+        
+        if (type == MXKAuthenticationTypeLogin)
+        {
+            self.emailAndTokenTextField.returnKeyType = UIReturnKeyDone;
+            self.displayNameTextField.hidden = YES;
+            
+            self.viewHeightConstraint.constant = self.displayNameTextField.frame.origin.y;
+        }
+        else
+        {
+            self.emailAndTokenTextField.returnKeyType = UIReturnKeyNext;
+            self.displayNameTextField.hidden = NO;
+            
+            self.viewHeightConstraint.constant = 122;
+        }
+        
+        return YES;
     }
-    return super.actualHeight;
+    
+    return NO;
 }
 
-- (BOOL)areAllRequiredFieldsFilled
+- (NSString*)validateParameters
 {
-    BOOL ret = [super areAllRequiredFieldsFilled];
+    NSString *errorMsg = [super validateParameters];
+    
+    if (!errorMsg)
+    {
+        if (!self.areAllRequiredFieldsSet)
+        {
+            errorMsg = [NSBundle mxk_localizedStringForKey:@"login_invalid_param"];
+        }
+    }
+    
+    return errorMsg;
+}
+
+- (BOOL)areAllRequiredFieldsSet
+{
+    BOOL ret = [super areAllRequiredFieldsSet];
     
     // Check required fields //FIXME what are required fields in this authentication flow?
     ret = (ret && self.userLoginTextField.text.length && self.emailAndTokenTextField.text.length);
+    
     return ret;
-}
-
-- (void)setAuthType:(MXKAuthenticationType)authType
-{
-    // Set initial layout
-    self.userLoginTextField.hidden = NO;
-    self.promptEmailTokenLabel.hidden = YES;
-    
-    if (authType == MXKAuthenticationTypeLogin)
-    {
-        self.emailAndTokenTextField.returnKeyType = UIReturnKeyDone;
-    }
-    else
-    {
-        self.emailAndTokenTextField.returnKeyType = UIReturnKeyNext;
-    }
-    
-    super.authType = authType;
 }
 
 - (void)dismissKeyboard
 {
     [self.userLoginTextField resignFirstResponder];
     [self.emailAndTokenTextField resignFirstResponder];
+    [self.displayNameTextField resignFirstResponder];
     
     [super dismissKeyboard];
 }
@@ -89,6 +114,13 @@
     self.promptEmailTokenLabel.hidden = NO;
     self.emailAndTokenTextField.placeholder = nil;
     self.emailAndTokenTextField.returnKeyType = UIReturnKeyDone;
+    
+    self.displayNameTextField.hidden = YES;
+}
+
+- (NSString*)userId
+{
+    return self.userLoginTextField.text;
 }
 
 #pragma mark UITextField delegate
@@ -100,11 +132,8 @@
         // "Done" key has been pressed
         [textField resignFirstResponder];
         
-        if (self.delegate && [self.delegate respondsToSelector:@selector(authInputsDoneKeyHasBeenPressed:)])
-        {
-            // Launch authentication now
-            [self.delegate authInputsDoneKeyHasBeenPressed:self];
-        }
+        // Launch authentication now
+        [self.delegate authInputsViewDidPressDoneKey:self];
     }
     else
     {
@@ -120,6 +149,48 @@
     }
     
     return YES;
+}
+
+#pragma mark -
+
+- (MXAuthenticationSession*)validateAuthenticationSession:(MXAuthenticationSession*)authSession
+{
+    // Check whether at least one of the listed flow is supported.
+    BOOL isSupported = NO;
+    
+    for (MXLoginFlow *loginFlow in authSession.flows)
+    {
+        // Check whether flow type is defined
+        if ([loginFlow.type isEqualToString:kMXLoginFlowTypeEmailCode])
+        {
+            isSupported = YES;
+            break;
+        }
+        else if (loginFlow.stages.count == 1 && [loginFlow.stages.firstObject isEqualToString:kMXLoginFlowTypeEmailCode])
+        {
+            isSupported = YES;
+            break;
+        }
+    }
+    
+    if (isSupported)
+    {
+        if (authSession.flows.count == 1)
+        {
+            // Return the original session.
+            return authSession;
+        }
+        else
+        {
+            // Keep only the supported flow.
+            MXAuthenticationSession *updatedAuthSession = [[MXAuthenticationSession alloc] init];
+            updatedAuthSession.session = authSession.session;
+            updatedAuthSession.params = authSession.params;
+            updatedAuthSession.flows = @[[MXLoginFlow modelFromJSON:@{@"stages":@[kMXLoginFlowTypeEmailCode]}]];
+        }
+    }
+    
+    return nil;
 }
 
 @end